PLG-140: Added additional stage for deleted products during import
